[refactor] folder structure
