+++ src/Divider/index.tsx
@@ -1,0 +1,6 @@
+import { Divider, DividerProps } from './Divider';
+import { withTheme } from '../config';
+
+export { Divider };
+export type { DividerProps };
+export default withTheme(Divider, 'Divider');
