+++ src/Switch/index.tsx
@@ -1,0 +1,6 @@
+import { Switch, SwitchProps } from './Switch';
+import { withTheme } from '../config';
+
+export { Switch };
+export type { SwitchProps };
+export default withTheme(Switch, 'Switch');
