+++ src/Text/Text.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import {
-  Text,
+  Text as NativeText,
   StyleSheet,
   Platform,
   TextProps as TextProperties,
   TextStyle,
   StyleProp,
 } from 'react-native';
-import { fonts, withTheme } from '../config';
+import { fonts } from '../config';
 import { patchWebProps, RneFunctionComponent } from '../helpers';
 import normalize from '../helpers/normalizeText';
 
@@ -23,24 +23,22 @@ export type TextProps = TextProperties & {
   h4Style?: StyleProp<TextStyle>;
 };
 
-const TextElement: RneFunctionComponent<TextProps> = (props) => {
-  const {
-    style,
-    theme,
-    children = '',
-    h1,
-    h2,
-    h3,
-    h4,
-    h1Style,
-    h2Style,
-    h3Style,
-    h4Style,
-    ...rest
-  } = props;
-
+const Text: RneFunctionComponent<TextProps> = ({
+  style = {},
+  h1 = false,
+  h2 = false,
+  h3 = false,
+  h4 = false,
+  h1Style = {},
+  h2Style = {},
+  h3Style = {},
+  h4Style = {},
+  children = '',
+  theme,
+  ...rest
+}) => {
   return (
-    <Text
+    <NativeText
       style={StyleSheet.flatten([
         {
           ...Platform.select({
@@ -60,20 +58,8 @@ const TextElement: RneFunctionComponent<TextProps> = (props) => {
       {...patchWebProps(rest)}
     >
       {children}
-    </Text>
+    </NativeText>
   );
-};
-
-TextElement.defaultProps = {
-  h1: false,
-  h2: false,
-  h3: false,
-  h4: false,
-  style: {},
-  h1Style: {},
-  h2Style: {},
-  h3Style: {},
-  h4Style: {},
 };
 
 const styles = StyleSheet.create({
@@ -86,5 +72,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export { TextElement };
-export default withTheme(TextElement, 'Text');
+export { Text };
