+++ src/Text/index.tsx
@@ -1,0 +1,6 @@
+import { Text, TextProps } from './Text';
+import { withTheme } from '../config';
+
+export { Text };
+export type { TextProps };
+export default withTheme(Text, 'Text');
