+++ src/card/CardFeaturedTitle.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Platform, StyleSheet, TextStyle } from 'react-native';
 import normalize from '../helpers/normalizeText';
 import { fonts, withTheme } from '../config';
-import Text, { TextProps } from '../text/Text';
+import Text, { TextProps } from '../Text';
 import { RneFunctionComponent } from '../helpers';
 
 const CardFeaturedTitle: RneFunctionComponent<TextProps> = ({
