+++ src/config/__tests__/makeStyles.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { create } from 'react-test-renderer';
 import { makeStyles, useTheme } from '../makeStyles';
-import Text from '../../text/Text';
+import Text from '../../Text';
 import { ThemeProvider } from '..';
 
 describe('useTheme()', () => {
