+++ src/config/theme.tsx
@@ -8,7 +8,7 @@ import { ButtonProps } from '../buttons/Button';
 import { ButtonGroupProps } from '../buttons/ButtonGroup';
 import { CardProps } from '../card/Card';
 import { CheckBoxProps } from '../checkbox/CheckBox';
-import { DividerProps } from '../divider/Divider';
+import { DividerProps } from '../Divider';
 import { HeaderProps } from '../header/Header';
 import { IconProps } from '../icons/Icon';
 import { ImageProps } from '../image/Image';
@@ -19,10 +19,10 @@ import { PricingCardProps } from '../pricing/PricingCard';
 import { SearchBarProps } from '../searchbar/SearchBar';
 import { SliderProps } from '../slider/Slider';
 import { SocialIconProps } from '../social/SocialIcon';
-import { TextProps } from '../text/Text';
+import { TextProps } from '../Text';
 import { TileProps } from '../tile/Tile';
 import { TooltipProps } from '../tooltip/Tooltip';
-import { SwitchProps } from '../switch/switch';
+import { SwitchProps } from '../Switch';
 import { ListItemAccordionProps } from '../list/ListItemAccordion';
 import { TabItemProps, TabProps } from '../tab/Tab';
 import { TabViewProps } from '../tab/TabView';
